Refactor a DB query = 10/10 pylint score
